justship/requirements/base.txt
Django==3.2.7
graphene-django==2.15.0
celery==5.1.2
redis==3.5.3
python-dotenv==0.19.0
django-graphql-jwt==0.3.4
django-cors-headers==3.7.0
pre-commit==2.15.0
